backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
python-multipart
openai
nltk
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
yake
pypdf
python-docx
